Reorganize folder structure slightly
Separate code & tests to their own folders, reassign module main entrypoint
